feat: add OrderStatus enum and OrderUpdate DTO for enhanced order management

+++ backend/java/product-service/src/main/java/com/Loqal/productservice/entity/ProductDTO.java
@@ -18,7 +18,7 @@ public record ProductDTO(
 
     double price,
 
-    long quantity,
+    int quantity,
 
     List<String> image_urls
 ){}
